requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
black==23.7.0
click==8.1.6
colorama==0.4.6
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
